[add] Alteração nas tabelas Stroop e Tol; Padronização dos resultados e integração dos métodos stroop e tol (#34975838).

+++ src/modules/tests/infra/typeorm/entities/Stroop.ts
@@ -54,10 +54,10 @@ class Stroop {
     @Column()
     part: string;
 
-    @Column("int", { array: true })
+    @Column("float", { array: true })
     xpos: number[];
 
-    @Column("int", { array: true })
+    @Column("float", { array: true })
     ypos: number[];
 
     @Column("int", { array: true })
